fixed sources for quick search and readme
