+++ backend/main.py
@@ -37,7 +37,7 @@ import database as db
 from council import stream_council_chat
 from events import EventBus
 import quick_search as qs_module
-from agents.chat import chat_stream_generate, TOOL_TEMPLATES, detect_template_family
+from agents.chat import chat_stream_generate
 from agents.personas import (
     ensure_philosopher_persona as _ensure_philosopher_persona,
     seed_all_defaults as _seed_all_defaults,
@@ -48,12 +48,11 @@ from agents.personas import (
 )
 import hf as hf_module
 import hyprfit
-from hf import parse_ollama_progress
 import rag
 import storage_diagnostics
-import connectors
 import comfyui
 import voice
+import model_management
 import model_providers
 from image_prompt_enhancer import normalize_enhancer_response
 from research import (
@@ -1097,74 +1096,6 @@ class KBCreate(BaseModel):
     name: str
     description: str = ""
 
-class ToolCreate(BaseModel):
-    name: str
-    description: str = ""
-    filename: str
-    code: str
-
-class ToolUpdate(BaseModel):
-    name: Optional[str] = None
-    description: Optional[str] = None
-    code: Optional[str] = None
-
-class MCPServerCreate(BaseModel):
-    name: str
-    transport: str = "stdio"
-    command: str = ""
-    url: str = ""
-    args: list[str] = []
-    env: dict = {}
-    headers: dict = {}
-    allow_private: bool = False
-    enabled: bool = True
-
-class MCPServerUpdate(BaseModel):
-    name: Optional[str] = None
-    transport: Optional[str] = None
-    command: Optional[str] = None
-    url: Optional[str] = None
-    args: Optional[list[str]] = None
-    env: Optional[dict] = None
-    headers: Optional[dict] = None
-    allow_private: Optional[bool] = None
-    enabled: Optional[bool] = None
-
-class OpenAPIConnectorCreate(BaseModel):
-    name: str
-    spec_url: str = ""
-    spec_json: Any = ""
-    base_url: str = ""
-    auth: dict = {}
-    headers: dict = {}
-    allow_private: bool = False
-    enabled: bool = True
-
-class OpenAPIConnectorUpdate(BaseModel):
-    name: Optional[str] = None
-    spec_url: Optional[str] = None
-    spec_json: Optional[Any] = None
-    base_url: Optional[str] = None
-    auth: Optional[dict] = None
-    headers: Optional[dict] = None
-    allow_private: Optional[bool] = None
-    enabled: Optional[bool] = None
-
-class ModelConfigCreate(BaseModel):
-    name: str
-    base_model: str
-    system_prompt: str = ""
-    tool_ids: list[str] = []
-    kb_ids: list[str] = []
-    parameters: dict = {}
-
-class ModelConfigUpdate(BaseModel):
-    name: Optional[str] = None
-    base_model: Optional[str] = None
-    system_prompt: Optional[str] = None
-    tool_ids: Optional[list[str]] = None
-    kb_ids: Optional[list[str]] = None
-    parameters: Optional[dict] = None
 
 class ConversationSearchRequest(BaseModel):
     query: str
@@ -1324,27 +1255,7 @@ async def hyprfit_rescan_hardware():
 
 
 async def _delete_ollama_model(model_name: str) -> dict:
-    """Delete a local Ollama model. Tries alternate HF name formats if needed."""
-    names_to_try = [model_name]
-    if not model_name.startswith("hf.co/") and "/" in model_name:
-        names_to_try.append(f"hf.co/{model_name}")
-    if model_name.startswith("hf.co/"):
-        names_to_try.append(model_name[len("hf.co/"):])
-    last_err = None
-    for name in names_to_try:
-        try:
-            r = await http.request("DELETE", f"{config.OLLAMA_URL}/api/delete", json={"name": name})
-            if r.status_code in (200, 204):
-                return {"status": "deleted", "model": model_name}
-            err_text = r.text[:400]
-            if "not found" in err_text.lower() and name != names_to_try[-1]:
-                continue
-            last_err = err_text
-        except Exception as e:
-            last_err = str(e)
-    if last_err and "not found" in last_err.lower():
-        return {"status": "deleted", "model": model_name, "note": "already removed from Ollama"}
-    raise HTTPException(502, f"Failed to delete model: {last_err}")
+    return await model_management.delete_ollama_model(http, model_name)
 
 
 @app.delete("/api/models")
@@ -3982,515 +3893,6 @@ async def purge_image_studio():
         "journal_cleared": journal_cleared,
         "note": note,
     }
-
-
-# ============================================================
-# TOOLS
-# ============================================================
-@app.get("/api/tools")
-async def list_tools():
-    return await db.get_tools()
-
-
-@app.post("/api/tools")
-async def create_tool(req: ToolCreate):
-    id = f"tool-{uuid.uuid4().hex[:12]}"
-    safe_name = os.path.basename(req.filename or "tool.py")
-    if not safe_name or safe_name != (req.filename or ""):
-        raise HTTPException(400, "Invalid filename")
-    filepath = os.path.abspath(os.path.join(config.TOOLS_DIR, safe_name))
-    tools_root = os.path.abspath(config.TOOLS_DIR)
-    if filepath != tools_root and not filepath.startswith(tools_root + os.sep):
-        raise HTTPException(400, "Invalid filename")
-
-    await db.create_tool(id, req.name, req.description, safe_name, req.code)
-    os.makedirs(config.TOOLS_DIR, exist_ok=True)
-    with open(filepath, "w") as f:
-        f.write(req.code)
-
-    return {"id": id, **req.model_dump(exclude={"filename"}), "filename": safe_name}
-
-
-@app.post("/api/tools/upload")
-async def upload_tool(file: UploadFile = File(...)):
-    """Upload a .py file as a tool."""
-    safe_name = os.path.basename(file.filename or "tool.py")
-    if not safe_name.endswith(".py"):
-        raise HTTPException(400, "Only .py files accepted")
-    filepath = os.path.join(config.TOOLS_DIR, safe_name)
-    if not os.path.abspath(filepath).startswith(os.path.abspath(config.TOOLS_DIR)):
-        raise HTTPException(400, "Invalid filename")
-
-    content = await file.read()
-    code = content.decode("utf-8")
-    name = safe_name.replace(".py", "")
-    id = f"tool-{uuid.uuid4().hex[:12]}"
-
-    with open(filepath, "w") as f:
-        f.write(code)
-
-    await db.create_tool(id, name, f"Uploaded: {safe_name}", safe_name, code)
-    return {"id": id, "name": name, "filename": safe_name, "code": code}
-
-
-@app.patch("/api/tools/{tool_id}")
-async def update_tool(tool_id: str, req: ToolUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    await db.update_tool(tool_id, **kwargs)
-    return {"status": "updated"}
-
-
-@app.delete("/api/tools/{tool_id}")
-async def delete_tool(tool_id: str):
-    await db.delete_tool(tool_id)
-    return {"status": "deleted"}
-
-
-@app.put("/api/tools/{tool_id}")
-async def update_tool_put(tool_id: str, req: ToolUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    if kwargs:
-        await db.update_tool(tool_id, **kwargs)
-    return {"status": "updated"}
-
-
-# ============================================================
-# CONNECTORS
-# ============================================================
-def _spec_json_to_text(value: Any) -> str:
-    if value is None:
-        return ""
-    if isinstance(value, str):
-        return value
-    return json.dumps(value)
-
-
-def _connector_tool_public(tl: dict) -> dict:
-    out = dict(tl)
-    out["name"] = tl.get("display_name") or tl.get("external_name") or tl.get("tool_name")
-    out["description"] = tl.get("description", "")
-    out["icon"] = "plug"
-    out["connector"] = True
-    return out
-
-
-@app.get("/api/connector-tools")
-async def list_connector_tools(enabled_only: bool = Query(True)):
-    tools = await db.get_connector_tools(enabled_only=enabled_only)
-    return [_connector_tool_public(t) for t in tools]
-
-
-@app.get("/api/mcp-servers")
-async def list_mcp_servers():
-    return await db.get_mcp_servers()
-
-
-@app.post("/api/mcp-servers")
-async def create_mcp_server(req: MCPServerCreate):
-    server_id = f"mcp-{uuid.uuid4().hex[:12]}"
-    transport = (req.transport or "stdio").strip().lower()
-    if transport not in {"stdio", "http", "streamable_http", "sse"}:
-        raise HTTPException(400, "transport must be stdio, http, streamable_http, or sse")
-    await db.create_mcp_server(
-        server_id,
-        req.name,
-        transport=transport,
-        command=req.command,
-        url=req.url,
-        args=req.args,
-        env=req.env,
-        headers=req.headers,
-        allow_private=req.allow_private,
-        enabled=req.enabled,
-    )
-    return await db.get_mcp_server(server_id)
-
-
-@app.patch("/api/mcp-servers/{server_id}")
-async def update_mcp_server(server_id: str, req: MCPServerUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    if "transport" in kwargs:
-        kwargs["transport"] = (kwargs["transport"] or "stdio").strip().lower()
-        if kwargs["transport"] not in {"stdio", "http", "streamable_http", "sse"}:
-            raise HTTPException(400, "transport must be stdio, http, streamable_http, or sse")
-    await db.update_mcp_server(server_id, **kwargs)
-    return await db.get_mcp_server(server_id)
-
-
-@app.delete("/api/mcp-servers/{server_id}")
-async def delete_mcp_server(server_id: str):
-    await db.delete_mcp_server(server_id)
-    return {"status": "deleted"}
-
-
-@app.post("/api/mcp-servers/{server_id}/discover")
-async def discover_mcp_server(server_id: str):
-    try:
-        return await connectors.discover_mcp_server(http, server_id)
-    except Exception as e:
-        raise HTTPException(400, str(e))
-
-
-@app.post("/api/mcp-servers/{server_id}/health")
-async def health_mcp_server(server_id: str):
-    try:
-        result = await connectors.discover_mcp_server(http, server_id)
-        return {"status": "ok", "tool_count": result.get("tool_count", 0)}
-    except Exception as e:
-        raise HTTPException(400, str(e))
-
-
-@app.get("/api/mcp-servers/{server_id}/tools")
-async def list_mcp_server_tools(server_id: str):
-    return [_connector_tool_public(t) for t in await db.get_connector_tools("mcp", server_id, enabled_only=False)]
-
-
-@app.get("/api/openapi-connectors")
-async def list_openapi_connectors():
-    return await db.get_openapi_connectors()
-
-
-@app.post("/api/openapi-connectors")
-async def create_openapi_connector(req: OpenAPIConnectorCreate):
-    connector_id = f"openapi-{uuid.uuid4().hex[:12]}"
-    await db.create_openapi_connector(
-        connector_id,
-        req.name,
-        spec_url=req.spec_url,
-        spec_json=_spec_json_to_text(req.spec_json),
-        base_url=req.base_url,
-        auth=req.auth,
-        headers=req.headers,
-        allow_private=req.allow_private,
-        enabled=req.enabled,
-    )
-    return await db.get_openapi_connector(connector_id)
-
-
-@app.patch("/api/openapi-connectors/{connector_id}")
-async def update_openapi_connector(connector_id: str, req: OpenAPIConnectorUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    if "spec_json" in kwargs:
-        kwargs["spec_json"] = _spec_json_to_text(kwargs["spec_json"])
-    await db.update_openapi_connector(connector_id, **kwargs)
-    return await db.get_openapi_connector(connector_id)
-
-
-@app.delete("/api/openapi-connectors/{connector_id}")
-async def delete_openapi_connector(connector_id: str):
-    await db.delete_openapi_connector(connector_id)
-    return {"status": "deleted"}
-
-
-@app.post("/api/openapi-connectors/{connector_id}/discover")
-async def discover_openapi_connector(connector_id: str):
-    try:
-        return await connectors.discover_openapi_connector(http, connector_id)
-    except Exception as e:
-        raise HTTPException(400, str(e))
-
-
-@app.post("/api/openapi-connectors/{connector_id}/health")
-async def health_openapi_connector(connector_id: str):
-    try:
-        result = await connectors.discover_openapi_connector(http, connector_id)
-        return {"status": "ok", "tool_count": result.get("tool_count", 0)}
-    except Exception as e:
-        raise HTTPException(400, str(e))
-
-
-@app.get("/api/openapi-connectors/{connector_id}/tools")
-async def list_openapi_connector_tools(connector_id: str):
-    return [_connector_tool_public(t) for t in await db.get_connector_tools("openapi", connector_id, enabled_only=False)]
-
-
-# ============================================================
-# MODEL CONFIGS
-# ============================================================
-@app.get("/api/model-configs")
-async def list_model_configs():
-    return await db.get_model_configs()
-
-
-@app.post("/api/model-configs")
-async def create_model_config(req: ModelConfigCreate):
-    id = f"mc-{uuid.uuid4().hex[:12]}"
-    await db.create_model_config(id, req.name, req.base_model, req.system_prompt, req.tool_ids, req.kb_ids, req.parameters)
-    return {"id": id, **req.model_dump()}
-
-
-@app.patch("/api/model-configs/{mc_id}")
-async def update_model_config(mc_id: str, req: ModelConfigUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    await db.update_model_config(mc_id, **kwargs)
-    return {"status": "updated"}
-
-
-@app.put("/api/model-configs/{mc_id}")
-async def update_model_config_put(mc_id: str, req: ModelConfigUpdate):
-    kwargs = {k: v for k, v in req.model_dump().items() if v is not None}
-    await db.update_model_config(mc_id, **kwargs)
-    return {"status": "updated"}
-
-
-@app.delete("/api/model-configs/{mc_id}")
-async def delete_model_config(mc_id: str):
-    await db.delete_model_config(mc_id)
-    return {"status": "deleted"}
-
-
-# ============================================================
-# OLLAMA MODEL MANAGEMENT
-# ============================================================
-@app.post("/api/models/pull")
-async def pull_model(request: Request):
-    """Pull a model from Ollama library — streams progress."""
-    body = await request.json()
-    model_name = body.get("name", "")
-    if not model_name:
-        raise HTTPException(400, "Model name required")
-
-    async def generate():
-        try:
-            got_done = False
-            async with http.stream("POST", f"{config.OLLAMA_URL}/api/pull",
-                                   json={"name": model_name, "stream": True},
-                                   timeout=httpx.Timeout(7200.0, connect=10.0)) as response:
-                if response.status_code != 200:
-                    err_body = (await response.aread()).decode()[:300]
-                    yield f"data: {json.dumps({'error': f'Ollama returned HTTP {response.status_code}: {err_body}'})}\n\n"
-                    return
-                async for line in response.aiter_lines():
-                    if not line.strip():
-                        continue
-                    sse, key = parse_ollama_progress(line, model_name)
-                    if not sse:
-                        continue
-                    if key == "error":
-                        yield sse
-                        return
-                    if key == "done":
-                        got_done = True
-                    yield sse
-            if not got_done:
-                # Stream ended without success — verify model exists
-                try:
-                    check = await http.post(f"{config.OLLAMA_URL}/api/show", json={"name": model_name})
-                    if check.status_code == 200:
-                        yield f"data: {json.dumps({'status': 'done', 'message': 'Pull complete'})}\n\n"
-                    else:
-                        yield f"data: {json.dumps({'error': 'Pull stream ended without confirmation'})}\n\n"
-                except Exception:
-                    yield f"data: {json.dumps({'error': 'Pull stream ended — could not verify model'})}\n\n"
-        except Exception as e:
-            yield f"data: {json.dumps({'error': str(e)})}\n\n"
-
-    return StreamingResponse(generate(), media_type="text/event-stream")
-
-
-@app.delete("/api/models/{model_name:path}")
-async def delete_model(model_name: str):
-    return await _delete_ollama_model(model_name)
-
-
-@app.post("/api/models/{model_name:path}/create-tool-model")
-async def create_tool_model(model_name: str):
-    """Patch an HF GGUF model's existing modelfile with a tool-calling TEMPLATE and save as a new model."""
-    import re as _re
-
-    try:
-        show_r = await http.post(f"{config.OLLAMA_URL}/api/show", json={"name": model_name, "verbose": True})
-        show_r.raise_for_status()
-        existing_mf = show_r.json().get("modelfile", "")
-    except Exception as e:
-        raise HTTPException(502, f"Could not fetch modelfile: {e}")
-
-    b = model_name.lower()
-
-    if any(x in b for x in ["qwen2.5", "qwen3", "qwen2"]):
-        template = (
-            "{{- if or .System .Tools }}<|im_start|>system\n"
-            "{{- if .System }}\n{{ .System }}\n{{- end }}\n"
-            "{{- if .Tools }}\n\n# Tools\n\nYou may call one or more functions to assist with the user query.\n\n"
-            "You are provided with function signatures within <tools></tools> XML tags:\n\n<tools>\n"
-            "{{- range .Tools }}\n{\"type\": \"function\", \"function\": {{ .Function }}}\n{{- end }}\n</tools>\n\n"
-            "For each function call, return a json object with function name and arguments within <tool_call></tool_call> XML tags:\n\n"
-            "<tool_call>\n{\"name\": <function-name>, \"arguments\": <args-json-object>}\n</tool_call>\n"
-            "{{- end }}<|im_end|>\n{{ end }}"
-            "{{- range .Messages }}"
-            "{{- if eq .Role \"user\" }}<|im_start|>user\n{{ .Content }}<|im_end|>\n"
-            "{{- else if eq .Role \"assistant\" }}<|im_start|>assistant\n"
-            "{{- if .Content }}{{ .Content }}"
-            "{{- else if .ToolCalls }}{{- range .ToolCalls }}<tool_call>\n{\"name\": \"{{ .Function.Name }}\", \"arguments\": {{ .Function.Arguments }}}\n</tool_call>\n{{- end }}"
-            "{{- end }}<|im_end|>\n"
-            "{{- else if eq .Role \"tool\" }}<|im_start|>user\n<tool_response>\n{{ .Content }}\n</tool_response><|im_end|>\n"
-            "{{- end }}{{- end }}<|im_start|>assistant\n"
-        )
-    elif any(x in b for x in ["llama-3", "llama3"]):
-        template = (
-            "{{- if or .System .Tools }}<|start_header_id|>system<|end_header_id|>\n\n"
-            "{{- if .System }}{{ .System }}\n{{ end }}"
-            "{{- if .Tools }}Environment: ipython\nTools: {{ .Tools }}\n{{ end }}"
-            "<|eot_id|>{{ end }}"
-            "{{- range .Messages }}"
-            "{{- if eq .Role \"user\" }}<|start_header_id|>user<|end_header_id|>\n\n{{ .Content }}<|eot_id|>"
-            "{{- else if eq .Role \"assistant\" }}<|start_header_id|>assistant<|end_header_id|>\n\n"
-            "{{- if .Content }}{{ .Content }}<|eot_id|>"
-            "{{- else if .ToolCalls }}<|python_tag|>{{ range .ToolCalls }}{\"name\": \"{{ .Function.Name }}\", \"parameters\": {{ .Function.Arguments }}}{{ end }}<|eot_id|>"
-            "{{- end }}"
-            "{{- else if eq .Role \"tool\" }}<|start_header_id|>ipython<|end_header_id|>\n\n{{ .Content }}<|eot_id|>"
-            "{{- end }}{{- end }}<|start_header_id|>assistant<|end_header_id|>\n\n"
-        )
-    elif any(x in b for x in ["mistral", "mixtral"]):
-        template = (
-            "[INST] {{- if .System }}{{ .System }}\n{{ end }}"
-            "{{- range .Messages }}{{- if eq .Role \"user\" }}{{ .Content }} [/INST] "
-            "{{- else if eq .Role \"assistant\" }}{{ .Content }}</s>[INST] "
-            "{{- else if eq .Role \"tool\" }}{{ .Content }} [/INST] "
-            "{{- end }}{{- end }}"
-        )
-    else:
-        template = (
-            "{{- if or .System .Tools }}<|im_start|>system\n"
-            "{{- if .System }}{{ .System }}\n{{- end }}"
-            "{{- if .Tools }}\nAvailable tools:\n{{- range .Tools }}\n{{ .Function }}\n{{- end }}\n{{- end }}"
-            "<|im_end|>\n{{ end }}"
-            "{{- range .Messages }}"
-            "{{- if eq .Role \"user\" }}<|im_start|>user\n{{ .Content }}<|im_end|>\n"
-            "{{- else if eq .Role \"assistant\" }}<|im_start|>assistant\n{{ .Content }}<|im_end|>\n"
-            "{{- else if eq .Role \"tool\" }}<|im_start|>tool\n{{ .Content }}<|im_end|>\n"
-            "{{- end }}{{- end }}<|im_start|>assistant\n"
-        )
-
-    from_match = _re.search(r'^# FROM (.+)$', existing_mf, _re.MULTILINE)
-    from_line = from_match.group(1).strip() if from_match else model_name
-
-    params = {}
-    for line in existing_mf.splitlines():
-        pm = _re.match(r'^PARAMETER\s+(\w+)\s+(.+)$', line.strip(), _re.IGNORECASE)
-        if pm:
-            key, val = pm.group(1).lower(), pm.group(2).strip()
-            try:
-                params[key] = float(val) if '.' in val else int(val)
-            except ValueError:
-                params[key] = val
-
-    payload: dict = {"name": model_name, "from": from_line, "template": template}
-    if params:
-        payload["parameters"] = params
-
-    try:
-        r = await http.post(
-            f"{config.OLLAMA_URL}/api/create",
-            json=payload,
-            timeout=120,
-        )
-        if r.status_code not in (200, 201):
-            raise HTTPException(r.status_code, f"Ollama error: {r.text[:400]}")
-        return {"status": "updated", "name": model_name}
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(502, f"Failed to create model: {e}")
-
-
-@app.get("/api/models/{model_name:path}/info")
-async def model_info(model_name: str):
-    """Get model details from Ollama."""
-    try:
-        r = await http.post(f"{config.OLLAMA_URL}/api/show", json={"name": model_name})
-        if r.status_code == 404:
-            raise HTTPException(404, f"Model '{model_name}' not found in Ollama")
-        r.raise_for_status()
-        return r.json()
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(502, f"Failed to get model info: {e}")
-
-
-@app.get("/api/models/{model_name:path}/template-info")
-async def get_template_info(model_name: str):
-    detected = detect_template_family(model_name)
-    return {
-        "detected": detected,
-        "templates": {k: {"label": v["label"]} for k, v in TOOL_TEMPLATES.items()},
-    }
-
-
-@app.post("/api/models/{model_name:path}/fix-template")
-async def fix_model_template(model_name: str, body: dict = Body(default={})):
-    """Patch a model's Modelfile to add a tool-calling template and recreate it in Ollama."""
-    family = body.get("family") or detect_template_family(model_name)
-    tpl = TOOL_TEMPLATES.get(family)
-    if not tpl:
-        raise HTTPException(400, f"Unknown template family: {family}")
-
-    stop_list = tpl["stops"]
-    create_payload = {
-        "model": model_name,
-        "from": model_name,
-        "template": tpl["template"],
-        "parameters": {"stop": stop_list},
-    }
-
-    try:
-        create_r = await http.post(
-            f"{config.OLLAMA_URL}/api/create",
-            json=create_payload,
-            timeout=120,
-        )
-        if create_r.status_code not in (200, 201):
-            raise HTTPException(502, f"Ollama create failed: {create_r.text[:300]}")
-    except HTTPException:
-        raise
-    except Exception as e:
-        raise HTTPException(502, f"Failed to recreate model: {e}")
-
-    return {"ok": True, "family": family, "model": model_name}
-
-
-# ============================================================
-# PERSONAS (Model Configs with avatars)
-# ============================================================
-@app.post("/api/model-configs/{mc_id}/avatar")
-async def upload_persona_avatar(mc_id: str, file: UploadFile = File(...)):
-    """Upload an avatar image for a persona/model config."""
-    existing = await db.get_model_config(mc_id)
-    if not existing:
-        raise HTTPException(404, "Model config not found")
-    avatar_dir = os.path.join(config.UPLOAD_DIR, "avatars")
-    os.makedirs(avatar_dir, exist_ok=True)
-
-    fname = file.filename or ""
-    raw_ext = fname.rsplit(".", 1)[-1].lower()[:10] if "." in fname else "png"
-    if raw_ext not in ("png", "jpg", "jpeg", "gif", "webp", "svg"):
-        raise HTTPException(400, "Invalid image type — allowed: png, jpg, jpeg, gif, webp, svg")
-    ext = raw_ext
-    avatar_path = os.path.join(avatar_dir, f"{mc_id}.{ext}")
-
-    content = await file.read()
-    with open(avatar_path, "wb") as f:
-        f.write(content)
-
-    params = existing.get("parameters", {}) if existing else {}
-    await db.update_model_config(mc_id, parameters={**params, "avatar": f"/api/avatars/{mc_id}.{ext}"})
-    return {"avatar_url": f"/api/avatars/{mc_id}.{ext}"}
-
-
-@app.get("/api/avatars/{filename}")
-async def get_avatar(filename: str):
-    """Serve avatar images."""
-    avatar_dir = os.path.join(config.UPLOAD_DIR, "avatars")
-    safe_name = os.path.basename(filename)
-    if not safe_name or safe_name != filename:
-        raise HTTPException(400, "Invalid filename")
-    filepath = os.path.join(avatar_dir, safe_name)
-    if not os.path.abspath(filepath).startswith(os.path.abspath(avatar_dir)):
-        raise HTTPException(400, "Invalid filename")
-    if not os.path.exists(filepath):
-        raise HTTPException(404, "Avatar not found")
-    return FileResponse(filepath)
 
 
 # ============================================================
