+++ deploy_monitor.py
@@ -53,6 +53,7 @@ WATCHED = {
     "backend/config.py":            ("Config",           REMOTE_BACKEND,            True),
     "backend/database.py":          ("Database",         REMOTE_BACKEND,            True),
     "backend/tools.py":             ("Tools",            REMOTE_BACKEND,            True),
+    "backend/model_management.py":  ("Model Management", REMOTE_BACKEND,            True),
     "backend/db/__init__.py":       ("Database Package", REMOTE_DB,                 True),
     "backend/db/schema.py":         ("Database Schema",  REMOTE_DB,                 True),
     "backend/routes/__init__.py":   ("Routes Package",   REMOTE_ROUTES,             True),
@@ -63,6 +64,9 @@ WATCHED = {
     "backend/routes/audio.py":      ("Audio Routes",     REMOTE_ROUTES,             True),
     "backend/routes/model_providers.py": ("Provider Routes", REMOTE_ROUTES,          True),
     "backend/routes/hf.py":         ("HF Routes",        REMOTE_ROUTES,             True),
+    "backend/routes/tools_connectors.py": ("Tools/Connector Routes", REMOTE_ROUTES,  True),
+    "backend/routes/model_configs.py": ("Model Config Routes", REMOTE_ROUTES,        True),
+    "backend/routes/ollama_models.py": ("Ollama Model Routes", REMOTE_ROUTES,        True),
     "backend/tooling/__init__.py":  ("Tooling Package",  REMOTE_TOOLING,            True),
     "backend/tooling/parser.py":    ("Tool Parser",      REMOTE_TOOLING,            True),
     "backend/connectors.py":        ("Connectors",       REMOTE_BACKEND,            True),
